+++ ocpmodels/datasets/__init__.py
@@ -11,4 +11,5 @@ from .lmdb_dataset import (
 )
 from .oc22_lmdb_dataset import OC22LmdbDataset
 
-from .ase_read_dataset import AseReadDataset
+from .ase_datasets import AseReadDataset
+from. ase_datasets import ASEDBDataset
