+++ ocpmodels/datasets/ase_datasets.py
@@ -1,0 +1,158 @@
+from pathlib import Path
+import ase
+import warnings
+
+from torch.utils.data import Dataset
+
+from ocpmodels.common.registry import registry
+from ocpmodels.preprocessing import AtomsToGraphs
+
+
+@registry.register_dataset("ase_read")
+class AseReadDataset(Dataset):
+    """
+    This Dataset uses ase.io.read to load data from a directory on disk.
+    This is intended for small-scale testing and demonstrations of OCP.
+    Larger datasets are better served by the efficiency of other dataset types
+    such as LMDB.
+
+    For a full list of ASE-readable filetypes, see
+    https://wiki.fysik.dtu.dk/ase/ase/io/io.html
+
+    args:
+        config (dict):
+            src (str): The source folder that contains your ASE-readable files
+
+            suffix (str): The ending of the filepath of each file you want to read
+                    ex. '/POSCAR', '.cif', '.xyz'
+
+            a2g_config (dict): configuration for ocp.preprocessing.AtomsToGraphs
+                    default options will work for most users
+                    
+                    If you are using this for a training dataset, set
+                    "r_energy"=True and/or "r_forces"=True as appropriate
+                    In that case, energy/forces must be in the files you read (ex. OUTCAR)
+
+        transform (callable, optional): Additional preprocessing function
+    """
+
+    def __init__(self, config, transform=None):
+        super(AseReadDataset, self).__init__()
+        self.config = config
+
+        self.path = Path(self.config["src"])
+        if self.path.is_file():
+            raise Exception("The specified src is not a directory")
+        self.id = sorted(self.path.glob(f'*{self.config["suffix"]}'))
+
+        a2g_config = config.get("a2g_config", {})
+        self.a2g = AtomsToGraphs(
+            max_neigh=a2g_config.get("max_neigh", 1000),
+            radius=a2g_config.get("radius", 8),
+            r_edges=a2g_config.get("r_edges", False),
+            r_energy=a2g_config.get("r_energy", False),
+            r_forces=a2g_config.get("r_forces", True),
+            r_distances=a2g_config.get("r_distances", True),
+            r_fixed=a2g_config.get("r_fixed", True),
+            r_pbc=a2g_config.get("r_pbc", True),
+        )
+
+        self.transform = transform
+
+    def __len__(self):
+        return len(self.id)
+
+    def __getitem__(self, idx):
+        try:
+            atoms = ase.io.read(self.id[idx])
+        except Exception as err:
+            warnings.warn(f"{err} occured for: {self.id[idx]}")
+
+        data_object = self.a2g.convert(atoms)
+
+        if self.transform is not None:
+            data_object = self.transform(data_object)
+
+        return data_object
+    
+    # This method is sometimes called by a trainer,
+    # but there is nothing necessary to do here
+    def close_db(self):
+        pass
+
+    
+@registry.register_dataset("ase_db")
+    class AseDBDataset(Dataset):
+    """
+    This Dataset connects to an ASE Database, allowing the storage of atoms objects
+    with a variety of backends including JSON, SQLite, and database server options.
+
+    For more information, see:
+    https://databases.fysik.dtu.dk/ase/ase/db/db.html
+
+    args:
+        config (dict):
+            src (str): The path to or connection address of your ASE DB
+
+            connect_args (dict): Additional arguments for ase.db.connect
+            
+            select_args (dict): Additional arguments for ase.db.select
+                    You can use this to query/filter your database
+
+            a2g_config (dict): configuration for ocp.preprocessing.AtomsToGraphs
+                    default options will work for most users
+                    
+                    If you are using this for a training dataset, set
+                    "r_energy"=True and/or "r_forces"=True as appropriate
+                    In that case, energy/forces must be in the database
+
+        transform (callable, optional): Additional preprocessing function
+    """
+    
+        def __init__(self, config, transform=None):
+            super(LmdbDataset, self).__init__()
+            self.config = config
+
+            self.db = self.connect_db(
+                self.config["src"], 
+                self.config.get("connect_args", {})
+            )
+            
+            self.select_args = self.config.get("select_args", {})
+            
+            self.id = [row.id for row in db.select(**self.select_args)]
+            
+            a2g_config = config.get("a2g_config", {})
+            self.a2g = AtomsToGraphs(
+                max_neigh=a2g_config.get("max_neigh", 1000),
+                radius=a2g_config.get("radius", 8),
+                r_edges=a2g_config.get("r_edges", True),
+                r_energy=a2g_config.get("r_energy", False),
+                r_forces=a2g_config.get("r_forces", False),
+                r_distances=a2g_config.get("r_distances", True),
+                r_fixed=a2g_config.get("r_fixed", True),
+                r_pbc=a2g_config.get("r_pbc", True),
+            )
+            
+            self.transform = transform
+            
+            
+        def __getitem__(self, idx):
+            atoms = self.db.get_atoms(self.id[idx])
+            
+            data_object = self.a2g.convert(atoms)
+
+            if self.transform is not None:
+                data_object = self.transform(data_object)
+
+            return data_object
+
+            
+        def __len__(self):
+            return len(self.id)
+            
+        def connect_db(self, address, connect_args = {}):
+            return ase.db.connect(address, **connect_args)
+            
+        def close_db(self):
+            pass
